requirements.txt
Django==2.2.10
djangorestframework==3.11.0
eth-account==0.4.0
pysha3==1.0.2
requests==2.22.0
